+++ pkg/unittest/lib/src/test_case.dart
@@ -60,9 +60,9 @@ class TestCase {
   Completer _testComplete;
 
   TestCase._internal(this.id, this.description, this._testFunction)
-      : currentGroup = _currentContext.fullName,
-        _setUp = _currentContext.testSetup,
-        _tearDown = _currentContext.testTeardown;
+      : currentGroup = _environment.currentContext.fullName,
+        _setUp = _environment.currentContext.testSetup,
+        _tearDown = _environment.currentContext.testTeardown;
 
   bool get isComplete => !enabled || result != null;